Simplify tldwatch to use unified provider approach
