--- MIGRATION_STEPS.md
@@ -1,82 +1,0 @@
-# Migration Steps: Transitioning to the New Architecture
-
-This document outlines the specific steps to complete the migration from the old provider architecture to the new consolidated architecture.
-
-## Current Status
-
-We've created new files with the new architecture alongside the existing files:
-
-- New base provider: `base_provider.py` (replacing `base.py`)
-- New request handler: `request_handler.py` (new)
-- New provider factory: `provider_factory.py` (new)
-- New provider registry: `provider_registry.py` (new)
-- New provider implementations: `openai_provider.py`, `anthropic_provider.py` (replacing `openai.py`, `anthropic.py`)
-- New chunking module: `chunking.py` (new)
-- New summarizer: `summarizer_new.py` (replacing `summarizer.py`)
-- New CLI: `main_new.py` (replacing `main.py`)
-
-## Migration Steps
-
-### Step 1: Complete Provider Implementations
-
-1. Implement the remaining providers as outlined in `IMPLEMENTATION_PLAN.md`:
-   - Google (Gemini) Provider
-   - Groq Provider
-   - Cerebras Provider
-   - DeepSeek Provider
-   - Ollama Provider
-
-### Step 2: Testing
-
-1. Create tests for the new architecture
-2. Ensure all tests pass with both the old and new implementations
-3. Verify that the new implementation works correctly with all providers
-
-### Step 3: Replace Old Files
-
-Once testing is complete, replace the old files with the new ones:
-
-1. Rename `base_provider.py` to `base.py` (replacing the old file)
-2. Rename `openai_provider.py` to `openai.py` (replacing the old file)
-3. Rename `anthropic_provider.py` to `anthropic.py` (replacing the old file)
-4. Create the new provider files with their final names (without the `_provider` suffix)
-5. Rename `summarizer_new.py` to `summarizer.py` (replacing the old file)
-6. Rename `main_new.py` to `main.py` (replacing the old file)
-7. Update `__init__.py` to use the new imports
-
-### Step 4: Update Package Structure
-
-1. Update the `__init__.py` files to expose the new API
-2. Ensure backward compatibility where possible
-3. Update any import statements in other parts of the codebase
-
-### Step 5: Documentation
-
-1. Update the main README.md with information from README_UPDATE.md
-2. Finalize the MIGRATION_GUIDE.md for users
-3. Remove temporary files like README_UPDATE.md and MIGRATION_STEPS.md
-
-## Backward Compatibility
-
-To maintain backward compatibility during the transition:
-
-1. The new `BaseProvider` class should support the same interface as the old one
-2. The new `Summarizer` class should accept the same parameters as the old one
-3. The CLI should support both old and new options
-
-## Timeline
-
-1. Complete provider implementations (1-2 days)
-2. Testing and validation (1-2 days)
-3. File replacement and package structure updates (1 day)
-4. Documentation updates (1 day)
-5. Final review and release (1 day)
-
-## After Migration
-
-Once the migration is complete:
-
-1. Remove any temporary files and documentation
-2. Update version number to reflect the major architectural change
-3. Create a release with detailed release notes
-4. Notify users of the new architecture and migration guide
