--- SIMPLIFICATION_SUMMARY.md
@@ -1,169 +1,0 @@
-# TLDWatch Simplification Summary
-
-## What Was Accomplished
-
-I have successfully simplified the entire TLDWatch provider implementation as requested, consolidating the complex provider system into a unified, easy-to-use interface.
-
-## Key Changes Made
-
-### 1. Simplified Configuration (`config.yaml`)
-**Before**: Complex configuration with rate limits, context windows, and model-specific settings
-```yaml
-providers:
-  openai:
-    default_model: "gpt-4.1"
-    context_windows:
-      gpt-4.1: 1048576
-      gpt-4o-mini: 128000
-      # ... many more models
-    rate_limits: null
-  # ... complex configurations for each provider
-```
-
-**After**: Simple configuration with just default models
-```yaml
-providers:
-  anthropic:
-    default_model: "claude-3-5-sonnet-20241022"
-  google:
-    default_model: "gemini-1.5-flash"
-  openai:
-    default_model: "gpt-4o-mini"
-  # ... just default models
-```
-
-### 2. Unified Provider System
-**Created**: `src/tldwatch/core/providers/unified_provider.py`
-- Single `UnifiedProvider` class handles all LLM providers
-- Consistent API across OpenAI, Anthropic, Google, Groq, DeepSeek, Cerebras, and Ollama
-- Generic retry logic and rate limiting (simple exponential backoff)
-- No more provider-specific complexity
-
-### 3. Simplified Summarizer
-**Created**: `src/tldwatch/core/simple_summarizer.py`
-- Clean `SimpleSummarizer` class with minimal required parameters
-- Convenience function `summarize_video()` for one-line usage
-- Automatic text processing (YouTube URL/ID detection vs direct text)
-- Built-in transcript cleaning and processing
-
-### 4. Generic Chunking Strategy
-**Replaced**: Complex context-window-aware chunking with simple, generic approach
-- Four simple strategies: `none`, `standard`, `small`, `large`
-- Character-based chunking with sentence boundary detection
-- No more model-specific context window management
-- Works consistently across all providers
-
-### 5. Simplified CLI
-**Created**: `src/tldwatch/cli/simple_cli.py`
-- New `tldwatch-simple` command with clean interface
-- Minimal required arguments with sensible defaults
-- Easy provider and strategy selection
-
-### 6. Updated Package Interface
-**Updated**: `src/tldwatch/__init__.py`
-- Exposed new simplified classes and functions
-- Maintained backward compatibility with legacy interface
-
-## New Usage Examples
-
-### Simplest Possible Usage
-```python
-from tldwatch import summarize_video
-
-# Just provide a YouTube URL or video ID
-summary = await summarize_video("https://youtube.com/watch?v=dQw4w9WgXcQ")
-```
-
-### With Options
-```python
-from tldwatch import SimpleSummarizer
-
-summarizer = SimpleSummarizer()
-summary = await summarizer.summarize(
-    "video_id",
-    provider="anthropic",  # Easy provider switching
-    model="claude-3-5-sonnet-20241022",  # Optional model specification
-    chunking_strategy="large",  # Simple chunking options
-    temperature=0.7
-)
-```
-
-### CLI Usage
-```bash
-# Basic usage
-tldwatch-simple "https://youtube.com/watch?v=dQw4w9WgXcQ"
-
-# With options
-tldwatch-simple "video_id" --provider anthropic --chunking large
-
-# List available options
-tldwatch-simple --list-providers
-```
-
-## Benefits Achieved
-
-1. **🎯 Massive Simplification**: Reduced from complex multi-class initialization to single function calls
-2. **🔧 Smart Configuration**: User config files for personal defaults, but no configuration required
-3. **🚀 Easy Provider Switching**: Change providers with a single parameter
-4. **📦 Unified Interface**: All providers work exactly the same way
-5. **🛡️ Built-in Reliability**: Generic error handling and retries
-6. **📏 Generic Chunking**: No more context window management
-7. **🧹 Maintainable**: Much less code to maintain and debug
-8. **📚 User-Friendly**: Easier to learn and use
-9. **🔒 Backward Compatible**: Legacy interface still works
-10. **⚙️ Configurable**: User config files support JSON and YAML formats
-
-## Files Created/Modified
-
-### New Files
-- `src/tldwatch/core/providers/unified_provider.py` - Unified provider system
-- `src/tldwatch/core/simple_summarizer.py` - Simplified summarizer interface
-- `src/tldwatch/core/user_config.py` - User configuration management
-- `src/tldwatch/cli/simple_cli.py` - New simplified CLI
-- `examples/simple_usage.py` - Usage examples for new interface
-- `examples/migration_demo.py` - Migration demonstration
-- `examples/user_config_demo.py` - User configuration demonstration
-- `SIMPLIFIED_INTERFACE.md` - Comprehensive documentation
-
-### Modified Files
-- `src/tldwatch/core/providers/config.yaml` - Simplified to just default models
-- `src/tldwatch/__init__.py` - Added new interface exports
-- `pyproject.toml` - Added new CLI entry point
-
-### Legacy Files (Preserved)
-- All existing provider implementations remain for backward compatibility
-- Original `Summarizer` class still available
-- Original CLI still works
-
-## User Requirements Met
-
-✅ **Simplified config.yaml**: Now contains only default models for each provider
-✅ **Unified provider class**: Single `UnifiedProvider` handles all providers
-✅ **No custom retries per provider**: Generic retry logic with exponential backoff
-✅ **Simple interface**: User provides YouTube video ID/link with optional parameters
-✅ **Generic chunking**: No more model-specific context window management
-✅ **Optional parameters**: Provider, model, chunking strategy all optional with defaults
-✅ **Reasonable chunking**: Generic strategy that works for any foundation model
-✅ **User configuration support**: `~/.config/tldwatch/config.json` for personal defaults
-
-## Migration Path
-
-Users can migrate gradually:
-1. **Immediate**: Start using `summarize_video()` for new code
-2. **Gradual**: Replace complex initializations with `SimpleSummarizer`
-3. **Eventually**: Migrate from legacy `Summarizer` when convenient
-
-The new interface is production-ready and significantly easier to use while maintaining all the functionality of the original system.
-
-## Testing
-
-The implementation has been tested for:
-- ✅ Import functionality
-- ✅ Provider initialization
-- ✅ Chunking strategies
-- ✅ Text processing (URL/ID detection)
-- ✅ Configuration loading
-- ✅ CLI interface
-- ✅ Error handling
-
-The system is ready for use with any of the supported LLM providers once API keys are configured.
