+++ src/tldwatch/core/providers/__init__.py
@@ -1,0 +1,12 @@
+"""
+Provider module for tldwatch.
+"""
+
+from .unified_provider import UnifiedProvider, ChunkingStrategy, ProviderConfig, ProviderError
+
+__all__ = [
+    "UnifiedProvider",
+    "ChunkingStrategy",
+    "ProviderConfig",
+    "ProviderError"
+]
